requirements.txt
wheel
numpy==1.24.3
PyGLM==2.7.0
glfw==2.5.9
PyOpenGL==3.1.6
PyOpenGL_accelerate==3.1.6